* commit 'aeb535a492dae71f0edc00d91a7f0934fbaa7351':
  test_vendor_lib: Add tests

+++ vendor_libs/test_vendor_lib/src/packet_stream.cc
@@ -78,7 +78,7 @@ bool PacketStream::SendEvent(const EventPacket& event, int fd) const {
   LOG_INFO(LOG_TAG, "Sending event with size: %zu octets",
            event.GetPacketSize());
 
-  if (!SendAll({event.GetType()}, 1, fd)) {
+  if (!SendAll({static_cast<uint8_t>(event.GetType())}, 1, fd)) {
     LOG_ERROR(LOG_TAG, "Error: Could not send event type.");
     return false;
   }
